Refactor u2f and remove fatal log (will exit the application) Fix linting
