+++ package.json
@@ -44,6 +44,7 @@
     "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
     "webpack-fix-style-only-entries": "0.4.0",
+    "whatwg-fetch": "3.0.0",
     "worker-loader": "2.0.0"
   },
   "devDependencies": {
